requirements.txt
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
bit==0.8.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
coincurve==18.0.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
djangorestframework==3.14.0
ecdsa==0.18.0
greenlet==2.0.2
idna==3.4
msgpack==1.0.5
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
